Read Description
Moved images into image folder, updated code to reflect this.

Alex/James: Looks like the bullets aren't hitting the planes because
your hitboxes are offset from the actual sprite. Odds are either the
sprite is not centered on the hitbox or the hitbox calculations aren't
matching the sprites (maybe check the differences in dimensions)